[FIX] Dynamic pivot: header alignment
Task: 5922279
